Add `WithStdBuildInfo` option to binary SBOM generator
The `WithStdBuildInfo` option can be used to provide the Go standard
library [debug.BuildInfo] to the SBOM generator, so that it doesn't
need to be read from disk. This can be useful if this step was already
performed.

Note that it's (currently) still required to have the Go binary
available on disk, because its hashes are calculated as part of the
SBOM generation.